Updated Gerber files for final fabrication
Note: According to the gerber standard, an empty solder mask file implies a complete solder mask across the entire layer.

Kicad uses an inverted display system for soldermask. What that means is that when viewed in Kicad, the parts highlighted in the soldermask layer are the parts where no soldermask will be applied.

More information on their FAQ section on their website.